test: handle bad debt socialisation in tests
there are some protocols that have bad debt socialisation. Aave doesn't so we don't have worry about it in the aaveWrapper. When there is bad debt in the lending protocol, some lending protocol socialise it and reduce the share price. In that case, the assumption is that some insurance fund (VII Finance or the lending protocol itself) will come and burn their tokens to make sure there is no bank run on the Vault wrapper contracts