feat: remove complicated maxDeposit and maxWithdraw logic
the underlying asset for the wrappers is the underlying vault or the aToken. We don't need to concern ourselves with supply caps or the current liquidity in the lending protocol when calculating maxDeposit and maxWithdraw amounts for a user. This is wrapper doesn't care about the underlying lending protocol. If the underlying protocol lets user mint the vault shares, that means user didn't violate supply caps of that lending protocol. After withdrawing underlying vault shares, if the underlying protocol lets user withdraw underlying assets, that means lending protocol had enough liquidity. The vault wrapper doesn't have to concern itself with any of it. This simplifies the codebase which good for security