Add periphery contracts
AssetToAssetSwapHook - A hook that takes care of conversion to and from vault wrappers. Users can provide raw asset and get the raw asset back

LiquidityHelper - Add liquidity to a pools with the yield harvesting hook. The helper contract takes care of the conversion to and from vault wrappers in this case as well